feat: Build CryptoStar Dapp on Ethereum
This commit includes all changes required to complete the project.

1) Addition of smart contract functions

A number of functions have been added to the smart contract:

* A name and a symbol are added to the starNotary tokens.

* A function lookUptokenIdToStarInfo is added, that looks up a stars using its corresponding token ID. It then returns the name of the star.

* A function exchangeStars is added, which makes it possible for two users to exchange their star tokens.

* A function to transfer a star is added, which transfers a star, given its token ID, from the address of the sender to the destination address.

2) Addition of supporting unit tests

A number of unit tests have been added:

* It is checked that the token name and token symbol are added properly.

* It is checked that indeed two users can exchange their stars.

* It is checked that star tokens can be transferred from the address of the sender to the destination address.

* It is checked that the information of a star is correctly obtained.

3) Deployment of the smart contract to the Rinkeby public test network

The file truffle-config.js is configured so that the contract can be deployed to Rinkeby.

4) Modification of the front end of the Dapp

A new front end function is written so that when the button "Look Up a Star" is clicked, the information of the star is shown.

5) README update

This step creates the README file that you are viewing right now.